src/app/pyproject.toml
#
# Warren app
#
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "warren"
description = "The visualization platform for your learning analytics"
readme = "README.md"
authors = [
  { name="Open FUN (France Universite Numerique)", email="fun.dev@fun-mooc.fr" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.7"
license = {file = "LICENSE.md"}
keywords = ["Analytics", "xAPI", "LRS", "LTI"]
dependencies = [
    "Django==4.2.2",
    "django-configurations==2.4.1",
    "django-lti-toolbox==1.0.1",
    "dockerflow==2022.8.0",
    "gunicorn==20.1.0",
    "psycopg2-binary==2.9.5",
    "sentry-sdk==1.13.0",
]
dynamic = ["version"]

[project.urls]
"Homepage" = "https://github.com/openfun/warren"
"Bug Tracker" = "https://github.com/openfun/warren/issues"

[project.optional-dependencies]
dev = [
    "black==23.1.0",
    "build==0.10.0",
    "factory-boy==3.2.1",
    "Faker==17.0.0",
    "ipdb==0.13.11",
    "ipython==8.10.0",
    "pytest==7.2.1",
    "pytest-cov==4.0.0",
]
ci = [
    "twine==4.0.2",
]

[tool.setuptools]
packages = ["warren", "apps"]

[tool.setuptools.dynamic]
version = { attr = "warren.__version__" }

# Third party packages configuration
[tool.pytest.ini_options]
addopts = "-v --cov-report term-missing --cov=apps"
python_files = [
    "test_*.py",
    "tests.py",
]
